[feature] Added SeleniumTestMixin
