+++ requirements-test.txt
@@ -1,3 +1,3 @@
 # For testing Dependency loaders
 openwisp_controller @ https://github.com/openwisp/openwisp-controller/tarball/master
-selenium~=4.9.1
+
